data-streamer/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
-e file:../pycommon#egg=pycommon  # via -r requirements.in
aiofiles==0.5.0           # via sanic
aiohttp==3.6.2            # via elasticsearch-async
async-timeout==3.0.1      # via aiohttp, elasticsearch-async
attrs==19.3.0             # via aiohttp
certifi==2020.6.20        # via elasticsearch, requests
chardet==3.0.4            # via aiohttp, requests
dnspython==1.16.0         # via pycommon
docker-pycreds==0.4.0     # via docker
docker==3.7.0             # via -r requirements.in
elasticsearch-async==6.2.0  # via -r requirements.in, pycommon
elasticsearch==7.8.0      # via elasticsearch-async
httptools==0.1.1          # via sanic
idna==2.7                 # via pycommon, requests, yarl
motor==2.0.0              # via -r requirements.in, pycommon
multidict==4.7.6          # via aiohttp, sanic, yarl
mypy-extensions==0.4.3    # via mypy
mypy==0.740               # via pycommon
pymongo==3.7.1            # via -r requirements.in, motor, pycommon
python-bsonjs==0.2.0      # via -r requirements.in
python-dateutil==2.7.5    # via pycommon
requests==2.24.0          # via docker
sanic==18.12.0            # via -r requirements.in
six==1.15.0               # via docker, docker-pycreds, python-dateutil, websocket-client
typed-ast==1.4.1          # via mypy
typing-extensions==3.7.4.2  # via mypy
ujson==1.35               # via -r requirements.in, sanic
urllib3==1.25.9           # via elasticsearch, requests
uvloop==0.14.0            # via pycommon, sanic
websocket-client==0.57.0  # via docker
websockets==6.0           # via sanic
yarl==1.4.2               # via aiohttp
